Add text_input_utils.dart
